move ot3-only fixture to the relevant test files
can't share it i guess
