--- api/tests/opentrons/hardware_control/conftest.py
@@ -1,19 +1,0 @@
-from typing import Iterator
-from mock import patch, AsyncMock
-import pytest
-
-from opentrons.hardware_control import ThreadManager
-from opentrons.hardware_control.ot3api import OT3API
-
-
-@pytest.fixture
-def mock_move_to(ot3_hardware: ThreadManager[OT3API]) -> Iterator[AsyncMock]:
-    with patch.object(
-        ot3_hardware.managed_obj,
-        "move_to",
-        AsyncMock(
-            spec=ot3_hardware.managed_obj.move_to,
-            wraps=ot3_hardware.managed_obj.move_to,
-        ),
-    ) as mock_move:
-        yield mock_move
