+++ api/tests/opentrons/hardware_control/test_ot3_api.py
@@ -11,6 +11,19 @@ from opentrons.hardware_control.ot3api import OT3API
 from opentrons.hardware_control import ThreadManager
 
 
+@pytest.fixture
+def mock_move_to(ot3_hardware: ThreadManager[OT3API]) -> Iterator[AsyncMock]:
+    with patch.object(
+        ot3_hardware.managed_obj,
+        "move_to",
+        AsyncMock(
+            spec=ot3_hardware.managed_obj.move_to,
+            wraps=ot3_hardware.managed_obj.move_to,
+        ),
+    ) as mock_move:
+        yield mock_move
+
+
 @pytest.mark.parametrize(
     "attached,load",
     (
