+++ api/tests/opentrons/hardware_control/test_ot3_calibration.py
@@ -18,6 +18,19 @@ from opentrons.types import Point
 
 
 @pytest.fixture
+def mock_move_to(ot3_hardware: ThreadManager[OT3API]) -> Iterator[AsyncMock]:
+    with patch.object(
+        ot3_hardware.managed_obj,
+        "move_to",
+        AsyncMock(
+            spec=ot3_hardware.managed_obj.move_to,
+            wraps=ot3_hardware.managed_obj.move_to,
+        ),
+    ) as mock_move:
+        yield mock_move
+
+
+@pytest.fixture
 def mock_capacitive_probe(ot3_hardware: ThreadManager[OT3API]) -> Iterator[AsyncMock]:
     with patch.object(
         ot3_hardware.managed_obj,
